+++ playground/example-github-async.R
@@ -1,8 +1,5 @@
 devtools::load_all()
 
-options(shinyOAuth.print_errors = TRUE)
-options(shinyOAuth.print_traceback = TRUE)
-
 library(shiny)
 library(mirai)