+++ playground/example-keycloak-private-key-jwt.R
@@ -49,9 +49,6 @@
 
 devtools::load_all()
 
-options(shinyOAuth.print_errors = TRUE)
-options(shinyOAuth.print_traceback = TRUE)
-
 library(shiny)
 
 # ---------- Config ----------