+++ playground/example-keycloak.R
@@ -6,9 +6,6 @@
 
 devtools::load_all()
 
-options(shinyOAuth.print_errors = TRUE)
-options(shinyOAuth.print_traceback = TRUE)
-
 library(shiny)
 
 # Configure provider and client